requirements.txt
# AIForex Combined Requirements
# ========================

# Core Web and API
# ---------------
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.5.3
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
jinja2>=3.1.2
websockets>=12.0
httpx>=0.27.0
requests>=2.31.0
aiohttp>=3.9.3

# Authentication
# -------------
python-jose>=3.3.0
pyjwt>=2.6.0
passlib>=1.7.4
bcrypt>=4.0.1

# Database
# --------
sqlalchemy>=2.0.28
alembic>=1.13.1
psycopg2-binary>=2.9.9
redis>=5.0.2
asyncpg>=0.29.0
databases>=0.8.0
ormar>=0.12.2
supabase>=2.13.0
qdrant-client>=1.1.1
pgvector>=0.2.5

# AI and ML
# --------
openai>=1.0.0
anthropic>=0.20.0
langchain>=0.1.12
langchain-community>=0.0.28
langchain-core>=0.1.29
langgraph>=0.0.32
pydantic-ai>=0.1.0
langchainhub>=0.1.14
tiktoken>=0.5.2
transformers>=4.39.0
torch>=2.0.0
tensorflow>=2.12.0
azure-ai-openai>=1.0.0

# Data Processing
# --------------
numpy>=1.26.4
scipy>=1.13.0
pandas>=2.2.2
scikit-learn>=1.4.1
statsmodels>=0.14.1
polars>=0.20.10
pyarrow>=14.0.2
yfinance>=0.2.36
TA-Lib>=0.6.3
arch>=5.3.0

# Visualization
# ------------
plotly>=5.15.0
matplotlib>=3.8.3
seaborn>=0.13.2

# Utilities
# --------
eventlet>=0.33.3
pyyaml>=6.0.1
loguru>=0.7.2
tenacity>=8.2.3
pytz>=2024.1
arrow>=1.3.0
click>=8.1.7
tqdm>=4.66.2
orjson>=3.9.15
uuid>=1.30

# Testing
# ------
pytest>=8.0.2
pytest-asyncio>=0.23.4
pytest-mock>=3.12.0
pytest-cov>=4.1.0
qodo-cover>=1.0.0
playwright>=1.43.0

# Monitoring
# ---------
prometheus-client>=0.18.1
sentry-sdk>=1.40.0

# Development Tools
# ---------------
black>=24.2.0
isort>=5.13.2
mypy>=1.7.1
flake8>=7.0.0
pre-commit>=3.6.2
docker>=7.0.0

# UI
# --
streamlit>=1.32.0 